Create module and complete functions.
