+++ .gitignore
@@ -1,5 +1,5 @@
 # Project files
-/logs
+app/logs
 .DS_Store
 
 # Byte-compiled / optimized / DLL files
